more cleanup for incremental
